FAST EXAME pronto, nescessita revisão final

+++ themes/intranet/plugin/fast-exames/admin/system/exames/index.php
@@ -17,17 +17,24 @@ if (!empty($action)):
 
             case "active":
                 $AdminExames->ExeStatus($toaction[1], 1);
-                WSErro("Exame <b>$exame->ex_descricao</b> ativo com sucesso!", WS_ACCEPT);
+                WSErro("Exame <b>$exame->ex_descricao</b> concluido com sucesso!", WS_ACCEPT);
                 break;
 
             case "inative":
                 $AdminExames->ExeStatus($toaction[1], 0);
-                WSErro("Exame <b>$exame->ex_descricao</b> desativado com sucesso!", WS_ACCEPT);
+                WSErro("Exame <b>$exame->ex_descricao</b> aberto com sucesso!", WS_ACCEPT);
                 break;
 
             case "delete":
-                if ($AdminExames->ExeDelete($toaction[1])):
+                if ($AdminExames->ExeCancelar($toaction[1], 1)):
                     WSErro("Exame <b>$exame->ex_descricao</b> deletado com sucesso!", WS_ACCEPT);
+                else:
+                    WSErro("Erro ao deletar", WS_ERROR);
+                endif;
+                break;
+            case "undelete":
+                if ($AdminExames->ExeCancelar($toaction[1], 0)):
+                    WSErro("Exame <b>$exame->ex_descricao</b> ativo com sucesso!", WS_ACCEPT);
                 else:
                     WSErro("Erro ao deletar", WS_ERROR);
                 endif;
@@ -47,13 +54,11 @@ $Pager = new Pager(FAST_INCLUDE . "admin/&exe=exames/index&page=");
 $Pager->ExePager($getPage, 15);
 
 
-$FeMaterial = new FeMaterial();
-$FeMaterial->Execute()->FullRead("SELECT * FROM fe_exames ORDER BY ex_status LIMIT :limit OFFSET :offset", "limit={$Pager->getLimit()}&offset={$Pager->getOffset()}", true);
-
 $FeExames = new FeExames();
-$FeExames->Execute()->findAll();
+$FeExames->Execute()->FullRead("SELECT * FROM fe_exames ORDER BY ex_cancelado,ex_status LIMIT :limit OFFSET :offset", "limit={$Pager->getLimit()}&offset={$Pager->getOffset()}", true);
 
 if (!$FeExames->Execute()->getResult()):
+    $Pager->ReturnPage();
     WSErro("Nenhum solicitação de alteraçao de exame encontrado!", WS_INFOR);
 else:
     ?>
@@ -76,7 +81,7 @@ else:
             foreach ($FeExames->Execute()->getResult() as $exames):
                 extract((array) $exames);
                 ?>    
-                <tr>
+                <tr <?= ($ex_cancelado ? "class=\"danger\"" : "") ?>>
                     <td><?= $ex_descricao; ?></td>
                     <td><?= $ex_minemonico; ?></td>
                     <td><?= $AdminExames->Setor($fe_setor_exec); ?></td>
@@ -89,11 +94,15 @@ else:
                         <ul class="post_actions plugin">
                             <li><a class="act_edit" href="<?= FAST_INCLUDE ?>admin/&exe=exames/update&examesId=<?= $ex_id; ?>#form" title="Editar">Editar</a></li>
                             <?php if (!$ex_status): ?>
-                                <li><a class="act_ative" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=active/<?= $ex_id; ?>#form" title="Ativar">Ativar</a></li>
+                                <li><a class="act_ative" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=active/<?= $ex_id; ?>#form" title="Concluir">Ativar</a></li>
                             <?php else: ?>
-                                <li><a class="act_inative" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=inative/<?= $ex_id; ?>#form" title="Inativar">Inativar</a></li>
+                                <li><a class="act_inative" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=inative/<?= $ex_id; ?>#form" title="Abrir">Inativar</a></li>
                             <?php endif; ?>
-                            <li><a class="act_delete" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=delete/<?= $ex_id; ?>#form" title="Excluir">Deletar</a></li>
+                            <?php if (!$ex_cancelado): ?>
+                                <li><a class="act_delete" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=delete/<?= $ex_id; ?>#form" title="Excluir">Deletar</a></li>
+                            <?php else: ?>
+                                <li><a class="act_accept" href="<?= FAST_INCLUDE ?>admin/&exe=exames/index&action=undelete/<?= $ex_id; ?>#form" title="Excluir">Deletar</a></li>
+                            <?php endif; ?>
                         </ul>
                     </td>
                 </tr>
@@ -102,5 +111,11 @@ else:
             ?>
         </tbody>
     </table>
-
 <?php endif; ?>
+
+<div class="row" id="form">
+    <?php
+    $Pager->ExePaginator("fe_exames");
+    echo $Pager->getPaginator();
+    ?>
+</div>
